move gl-matrix from modules to libs
