+++ js/modules/bounding-box.js
@@ -1,4 +1,4 @@
-import './gl-matrix-min.js';
+import '/js/libs/gl-matrix-min.js';
 import { EYE } from './shader.js';
 
 const glMatrix = window.glMatrix;
