+++ js/modules/rubiks.js
@@ -1,4 +1,4 @@
-import './gl-matrix-min.js';
+import '/js/libs/gl-matrix-min.js';
 import BoundingBox from './bounding-box.js';
 import { EYE } from './shader.js';
